CORE-16208 add integration test for the mediator subscription
