+++ settings.gradle
@@ -453,6 +453,7 @@ include 'testing:sandboxes:test-api'
 include 'testing:sandboxes-testkit'
 include 'testing:security-manager-utilities'
 include 'testing:test-serialization'
+include 'testing:test-state-manager'
 include 'testing:test-utilities'
 include 'testing:uniqueness:backing-store-fake'
 include 'testing:uniqueness:uniqueness-utilities'
