+++ testing/message-patterns/test.db.bndrun
@@ -31,6 +31,7 @@
     bnd.identity;id='net.corda.db-topic-admin-impl',\
     bnd.identity;id='net.corda.db-message-bus-impl',\
     bnd.identity;id='net.corda.db-message-bus-testkit',\
+    bnd.identity;id='net.corda.test-state-manager',\
     bnd.identity;id='net.bytebuddy.byte-buddy',\
     bnd.identity;id='junit-jupiter-engine',\
     bnd.identity;id='junit-platform-launcher',\
