+++ testing/message-patterns/test.kafka.bndrun
@@ -29,6 +29,7 @@
     bnd.identity;id='net.corda.schema-registry-impl',\
     bnd.identity;id='net.corda.kafka-topic-admin-impl',\
     bnd.identity;id='net.corda.lifecycle-impl',\
+    bnd.identity;id='net.corda.test-state-manager',\
     bnd.identity;id='junit-jupiter-engine',\
     bnd.identity;id='junit-platform-launcher',\
     bnd.identity;id='slf4j.simple',\
