+++ testing/test-state-manager/src/main/java/net/corda/testing/statemanager/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.testing.statemanager;
+
+import org.osgi.annotation.bundle.Export;
